fix: decompose and refactor pagination
