+++ src/components/pagination-numbers/style.css
@@ -1,0 +1,32 @@
+.Pagination-list {
+    display: flex;
+    justify-content: flex-end;
+    gap: 2px;
+    list-style: none;
+}
+
+.Pagination-item {
+    width: 22px;
+    height: 22px;
+    text-align: center;
+    border-radius: 2px;
+    cursor: pointer;
+
+    &:hover {
+        background-color: #6B4ACB1A;
+    }
+}
+
+.Pagination-item_selected {
+    width: 22px;
+    height: 22px;
+    text-align: center;
+    border-radius: 2px;
+    background-color: var(--primary);
+    color: var(--second-text);
+}
+
+.Pagination-item_dots {
+    cursor: default;
+    color: #888;
+}
